+++ common/src/YM2413.c
@@ -1,6 +1,7 @@
 /*
  * YM2413: Master System FM
  * */
+#include <stdint.h>
 #include "Enums.h"
 #include "SoundSystem.h"
 
@@ -8,4 +9,4 @@ void stopSounds(void) {}
 
 void soundTick(void) {}
 
-void playSound(int ununused) {}
+void playSound(uint8_t ununused) {}
